requirements.txt
prompt_toolkit==3.0.51
playwright==1.42.0
parso==0.8.4
executing==2.2.0
jedi==0.19.2
pyee==11.0.1
pytest
pexpect
pillow>=9.0.0
pytest>=7.0.0
pure_eval==0.2.3
traitlets==5.14.3
cairosvg>=2.7.0
Jinja2==3.1.6
fastapi
ipython==9.2.0
cairosvg
Pygments==2.19.1
matplotlib-inline==0.1.7
networkx==3.5
typing_extensions==4.13.2
jsonpickle==4.1.0
ipython_pygments_lexers
wcwidth==0.2.13
MarkupSafe==3.0.2
stack-data==0.6.3
decorator==5.2.1
pyvis==0.3.2
click==8.2.1
greenlet
mermaid_cli==0.1.2
colorama
gitpython
uvicorn
openai
ptyprocess
asttokens==3.0.0
python-dotenv
pylint
